requirements.txt
PySide6>=6.0.0
openai>=1.0.0
pyinstaller>=6.0.0
python-docx
python-pptx
openpyxl
pypdf
pandas
duckduckgo-search
trafilatura
beautifulsoup4
requests
